Extract processor into a standalone file
